Added the SQL database object
This just abstracts a bunch of stuff away so that it is harder to screw
up. Some things I'll have to validate because the sqlite3 library isn't
the best, but it should work fine. There is a method on the object that
I am going to remove called get_ignored_embed eventually because after
thinking about it I want in in the main code instead of the database
code. It will stay for now so I can just copy and paste it later. Also
other random little things.